+++ changelog.md
@@ -1,6 +1,8 @@
 ### v2.2.0
 
 -   Feature: Add results page.
+-   Feature: Make header responsive.
+-   Feature: Make home banner responsive.
 
 ### v2.1.0